+++ python/ray/data/tests/test_tensor.py
@@ -1009,116 +1009,6 @@ def test_ragged_tensors(ray_start_regular_shared, restore_data_context, tensor_f
     ]
 
 
-def test_tensor_type_equality_checks():
-    # Test that different types are not equal
-    fs_tensor_type_v1 = ArrowTensorType((2, 3), pa.int64())
-    fs_tensor_type_v2 = ArrowTensorTypeV2((2, 3), pa.int64())
-
-    assert fs_tensor_type_v1 != fs_tensor_type_v2
-
-    # Test different shapes/dtypes aren't equal
-    assert fs_tensor_type_v1 != ArrowTensorType((3, 3), pa.int64())
-    assert fs_tensor_type_v1 != ArrowTensorType((2, 3), pa.float64())
-    assert fs_tensor_type_v2 != ArrowTensorTypeV2((3, 3), pa.int64())
-    assert fs_tensor_type_v2 != ArrowTensorTypeV2((2, 3), pa.float64())
-
-    # Test var-shaped tensor type
-    vs_tensor_type = ArrowVariableShapedTensorType(pa.int64(), 2)
-
-    # Test that different types are not equal
-    assert vs_tensor_type != ArrowVariableShapedTensorType(pa.int64(), 3)
-    assert vs_tensor_type != ArrowVariableShapedTensorType(pa.float64(), 2)
-    assert vs_tensor_type != fs_tensor_type_v1
-    assert vs_tensor_type != fs_tensor_type_v2
-
-
-def test_arrow_fixed_shape_tensor_type_eq_with_concat(restore_data_context):
-    """Test that ArrowTensorType and ArrowTensorTypeV2 __eq__ methods work correctly
-    when concatenating Arrow arrays with the same tensor type."""
-    from ray.data.context import DataContext
-    from ray.data.extensions.tensor_extension import (
-        ArrowTensorArray,
-        ArrowTensorType,
-        ArrowTensorTypeV2,
-    )
-
-    # Test ArrowTensorType V1
-    tensor_type_v1 = ArrowTensorType((2, 3), pa.int64())
-
-    DataContext.get_current().use_arrow_tensor_v2 = False
-    first = ArrowTensorArray.from_numpy(np.ones((2, 2, 3), dtype=np.int64))
-    second = ArrowTensorArray.from_numpy(np.zeros((3, 2, 3), dtype=np.int64))
-
-    assert first.type == second.type
-    # Assert commutation
-    assert tensor_type_v1 == first.type
-    assert first.type == tensor_type_v1
-
-    # Test concatenation works appropriately
-    concatenated = pa.concat_arrays([first, second])
-    assert len(concatenated) == 5
-    assert concatenated.type == tensor_type_v1
-
-    expected = np.vstack([first.to_numpy(), second.to_numpy()])
-    np.testing.assert_array_equal(concatenated.to_numpy(), expected)
-
-    # Test ArrowTensorTypeV2
-    tensor_type_v2 = ArrowTensorTypeV2((2, 3), pa.int64())
-
-    DataContext.get_current().use_arrow_tensor_v2 = True
-
-    first = ArrowTensorArray.from_numpy(np.ones((2, 2, 3), dtype=np.int64))
-    second = ArrowTensorArray.from_numpy(np.ones((3, 2, 3), dtype=np.int64))
-
-    assert first.type == second.type
-    # Assert commutation
-    assert tensor_type_v2 == first.type
-    assert first.type == tensor_type_v2
-
-    # Test concatenation works appropriately
-    concatenated_v2 = pa.concat_arrays([first, second])
-    assert len(concatenated_v2) == 5
-    assert concatenated_v2.type == tensor_type_v2
-
-    # Assert on the full concatenated array
-    expected = np.vstack([first.to_numpy(), second.to_numpy()])
-    np.testing.assert_array_equal(concatenated_v2.to_numpy(), expected)
-
-
-def test_arrow_variable_shaped_tensor_type_eq_with_concat():
-    """Test that ArrowVariableShapedTensorType __eq__ method works correctly
-    when concatenating Arrow arrays with variable shaped tensors."""
-    from ray.data.extensions.tensor_extension import (
-        ArrowVariableShapedTensorArray,
-        ArrowVariableShapedTensorType,
-    )
-
-    # Create ArrowVariableShapedTensorType
-    var_tensor_type = ArrowVariableShapedTensorType(pa.int64(), 2)
-
-    # Create arrays with variable-shaped tensors
-    tensors1 = [np.array([[1, 2], [3, 4]]), np.array([[5, 6, 7], [8, 9, 10]])]
-    tensors2 = [np.array([[11, 12, 13, 14]]), np.array([[15], [16], [17]])]
-
-    arr1 = ArrowVariableShapedTensorArray.from_numpy(tensors1)
-    arr2 = ArrowVariableShapedTensorArray.from_numpy(tensors2)
-
-    assert arr1.type == arr2.type
-    # Assert commutation
-    assert var_tensor_type == arr1.type
-    assert arr1.type == var_tensor_type
-
-    # Test concatenation works appropriately
-    concatenated = pa.concat_arrays([arr1, arr2])
-    assert len(concatenated) == 4
-    assert concatenated.type == var_tensor_type
-
-    result = concatenated.to_numpy()
-    expected_shapes = [(2, 2), (2, 3), (1, 4), (3, 1)]
-    for i, expected_shape in enumerate(expected_shapes):
-        assert result[i].shape == expected_shape
-
-
 if __name__ == "__main__":
     import sys
